+++ crates/burn-import/onnx-tests/tests/maxpool/mod.rs
@@ -1,0 +1,64 @@
+// Import the shared macro
+use crate::include_models;
+include_models!(maxpool1d, maxpool2d);
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use burn::tensor::{Tensor, TensorData};
+
+    type Backend = burn_ndarray::NdArray<f32>;
+
+    #[test]
+    fn maxpool1d() {
+        let device = Default::default();
+
+        let model: maxpool1d::Model<Backend> = maxpool1d::Model::new(&device);
+        let input = Tensor::<Backend, 3>::from_floats(
+            [[
+                [1.927, 1.487, 0.901, -2.106, 0.678],
+                [-1.235, -0.043, -1.605, -0.752, -0.687],
+                [-0.493, 0.241, -1.111, 0.092, -2.317],
+                [-0.217, -1.385, -0.396, 0.803, -0.622],
+                [-0.592, -0.063, -0.829, 0.331, -1.558],
+            ]],
+            &device,
+        );
+        let output = model.forward(input);
+        let expected = TensorData::from([[
+            [1.927f32, 1.927, 0.901],
+            [-0.043, -0.043, -0.687],
+            [0.241, 0.241, 0.092],
+            [-0.217, 0.803, 0.803],
+            [-0.063, 0.331, 0.331],
+        ]]);
+        output.to_data().assert_eq(&expected, true);
+    }
+
+    #[test]
+    fn maxpool2d() {
+        // Initialize the model without weights (because the exported file does not contain them)
+        let device = Default::default();
+        let model: maxpool2d::Model<Backend> = maxpool2d::Model::new(&device);
+
+        // Run the model
+        let input = Tensor::<Backend, 4>::from_floats(
+            [[[
+                [1.927, 1.487, 0.901, -2.106, 0.678],
+                [-1.235, -0.043, -1.605, -0.752, -0.687],
+                [-0.493, 0.241, -1.111, 0.092, -2.317],
+                [-0.217, -1.385, -0.396, 0.803, -0.622],
+                [-0.592, -0.063, -0.829, 0.331, -1.558],
+            ]]],
+            &device,
+        );
+        let output = model.forward(input);
+        let expected = TensorData::from([[[
+            [0.901f32, 1.927, 1.487, 0.901],
+            [0.901, 1.927, 1.487, 0.901],
+            [-0.396, 0.803, 0.241, -0.396],
+        ]]]);
+
+        output.to_data().assert_eq(&expected, true);
+    }
+}
