+++ crates/burn-import/onnx-tests/tests/test_mod.rs
@@ -8,11 +8,17 @@ mod node_tests;
 // Import individual node modules
 pub mod add;
 pub mod argmax;
+pub mod avg_pool;
+pub mod batch_norm;
+pub mod cast;
+pub mod clip;
 pub mod concat;
 pub mod constant;
 pub mod constant_of_shape;
 pub mod conv;
 pub mod conv_transpose;
+pub mod cos;
+pub mod cosh;
 pub mod div;
 pub mod dropout;
 pub mod erf;
@@ -21,9 +27,12 @@ pub mod global_avr_pool;
 pub mod log_softmax;
 pub mod matmul;
 pub mod max;
+pub mod maxpool;
 pub mod mean;
 pub mod min;
 pub mod mul;
+pub mod sin;
+pub mod sinh;
 pub mod slice;
 pub mod softmax;
 pub mod sqrt;
